+++ internal/app/dto/withdrawal.go
@@ -4,17 +4,17 @@ import "time"
 
 type Withdrawal struct {
 	OrderNum    string    `json:"order"`
-	Amount      int       `json:"sum"`
+	Amount      float32   `json:"sum"`
 	Status      string    `json:"status"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
 
 type Withdraw struct {
-	OrderNum string `json:"order"`
-	Amount   int    `json:"sum"`
+	OrderNum string  `json:"order"`
+	Amount   float32 `json:"sum"`
 }
 
 type Balance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
